Add option GenerateDebugInfo=<level>
