+++ substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/elf/dwarf/DirEntry.java
@@ -27,8 +27,15 @@
 package com.oracle.objectfile.elf.dwarf;
 
 /**
- * track the directory associated with one or
+ * Tracks the directory associated with one or
  * more source files.
+ *
+ * This is identified separately from each FileEntry
+ * idenityfing files that reside in the directory.
+ * That is necessary because the line info generator
+ * needs to collect and write out directory names
+ * into directory tables once only rather than once
+ * per file.
  */
 public class DirEntry {
     private String path;
